+++ src/estreeToString.js
@@ -1,5 +1,6 @@
 import {
-  visitorKeys as jsdocTypePrattParserVisitorKeys
+  visitorKeys as jsdocTypePrattParserVisitorKeys,
+  stringify
 } from 'jsdoc-type-pratt-parser';
 
 import {jsdocVisitorKeys} from './commentParserToESTree.js';
@@ -7,7 +8,7 @@ import {jsdocVisitorKeys} from './commentParserToESTree.js';
 const stringifiers = {
   JsdocBlock ({
     delimiter, postDelimiter, lineEnd, initial, terminal, endLine
-  }, descriptionLines, tags) {
+  }, opts, descriptionLines, tags) {
     const alreadyHasLine =
         (descriptionLines.length && !tags.length &&
           descriptionLines[descriptionLines.length - 1].endsWith('\n')) ||
@@ -36,11 +37,11 @@ const stringifiers = {
     return `${initial}${delimiter}${postDelimiter}${description}`;
   },
   JsdocTypeLine ({
-    initial, delimiter, postDelimiter, rawType
+    initial, delimiter, postDelimiter, rawType, parsedType
   }) {
     return `${initial}${delimiter}${postDelimiter}${rawType}`;
   },
-  JsdocTag (node, parsedType, typeLines, descriptionLines) {
+  JsdocTag (node, opts, parsedType, typeLines, descriptionLines) {
     const {
       description,
       name, postName, postTag, postType,
@@ -54,7 +55,9 @@ const stringifiers = {
       // parsedType
       // Comment this out later in favor of `parsedType`
       // We can't use raw `typeLines` as first argument has delimiter on it
-      typeLines.length ? `{${typeLines.join('\n')}}` : ''
+      (opts.preferRawType || !parsedType)
+        ? typeLines.length ? `{${typeLines.join('\n')}}` : ''
+        : parsedType
     }${postType}${
       name ? `${name}${postName || (description ? '\n' : '')}` : ''
     }${descriptionLines.join('\n')}`;
@@ -67,29 +70,29 @@ const visitorKeys = {...jsdocVisitorKeys, ...jsdocTypePrattParserVisitorKeys};
  * @todo convert for use by escodegen (until may be patched to support
  *   custom entries?).
  * @param {Node} node
+ * @param {{preferRawType: boolean}} opts
  * @throws {Error}
  * @returns {string}
  */
-function estreeToString (node) {
+function estreeToString (node, opts = {}) {
   if (Object.prototype.hasOwnProperty.call(stringifiers, node.type)) {
     const childNodeOrArray = visitorKeys[node.type];
 
     const args = childNodeOrArray.map((key) => {
       return Array.isArray(node[key])
         ? node[key].map((item) => {
-          return estreeToString(item);
+          return estreeToString(item, opts);
         })
         : (node[key] === undefined || node[key] === null
           ? []
-          : [estreeToString(node[key])]);
+          : [estreeToString(node[key], opts)]);
     });
-
-    return stringifiers[node.type](node, ...args);
+    return stringifiers[node.type](node, opts, ...args);
   }
 
   // We use raw type instead but it is a key as other apps may wish to traverse
   if (node.type.startsWith('JsdocType')) {
-    return '';
+    return opts.preferRawType ? '' : `{${stringify(node)}}`;
   }
 
   throw new Error(`Unhandled node type: ${node.type}`);